# backend/internal/apidocs/openapi.yaml
openapi: 3.1.0
info:
  title: Touchline API
  version: 0.1.0
  description: |
    Touchline is a text-first football management game. The backend is a single
    binary (`cmd/touchline serve` runs the API, the scheduler world clock, and
    the event worker in one process; the api/scheduler/worker subcommands split
    them for prod isolation).

    Authentication is cookie-based: login exchanges credentials for a rotating
    access/refresh cookie pair, and every game endpoint below a security scheme
    marker requires a valid `access_token` cookie. The whole world is scoped to
    the session, so "my world" reads are always the caller's own world.

    Launch model: worlds are created as metadata (name), then an admin declares
    countries and leagues, and finally seeds the whole world with clubs,
    players, and memberships. Seasons/league play follow the fixture system
    driven by the scheduler.

    The interactive reference is this page; the raw spec lives at
    `/openapi.yaml`. Live game state (world ticks, match pace 15-min ticks)
    streams over `/ws` (see the WebSocket section).
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development (cmd/touchline serve)
...
paths:
...
  /api/managers/me/policies/{type}:
    get:
      tags: [policy]
      summary: Get my delegation policy
      operationId: getPolicy
      description: Returns the manager's saved policy for the given type (squad, transfer, or training). 404 if none is saved.
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string, enum: [squad, transfer, training] }
      responses:
        "200":
          description: The saved policy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "404":
          description: No policy saved for this type.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    put:
      tags: [policy]
      summary: Save or replace my delegation policy
      operationId: upsertPolicy
      description: Creates or replaces the manager's policy for the given type. A saved policy overrides the assistant default; deleting it restores the default.
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string, enum: [squad, transfer, training] }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [params]
              properties:
                params: { type: object }
      responses:
        "200":
          description: The persisted policy.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Policy" }
        "400":
          description: Invalid policy type or params.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [policy]
      summary: Delete my delegation policy
      operationId: deletePolicy
      description: Removes a saved policy row; subsequent delegated actions fall back to the assistant default. Deleting an absent row is a no-op (404).
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string, enum: [squad, transfer, training] }
      responses:
        "204":
          description: Policy deleted.
        "404":
          description: No policy was saved for this type.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/managers/me/absence:
    get:
      tags: [policy]
      summary: My absence state
      operationId: getAbsence
      description: The manager's current away-mode status, associated clubs and the next scheduled fixture. The delegation_active flag is true when the policybot is currently acting on the manager's clubs.
      responses:
        "200":
          description: The absence view.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AbsenceView" }
    put:
      tags: [policy]
      summary: Toggle explicit away mode
      operationId: setAway
      description: Turns explicit away mode on or off. Toggling on activates delegation immediately and clears the missed-fixture streak. Toggling off reverses delegation until the streak rebuilds. The auto-away flag is always cleared on explicit toggle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [away]
              properties:
                away: { type: boolean }
      responses:
        "204":
          description: Away mode updated.
        "400":
          description: Missing or invalid away payload.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
    delete:
      tags: [policy]
      summary: Clear away mode
      operationId: clearAway
      description: Convenience alias for PUT with away=false.
      responses:
        "204":
          description: Away mode cleared.
...
  /api/matches/{id}/events:
    get:
      tags: [match]
      summary: Persisted match event feed
      operationId: getMatchEvents
      description: The full ordered match_events feed (the same list the live socket streams), world-scoped.
      parameters:
        - $ref: "#/components/parameters/MatchID"
      responses:
        "200":
          description: Ordered event feed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  events:
                    type: array
                    items: { $ref: "#/components/schemas/MatchEvent" }
              examples:
                sample:
                  value:
                    events:
                      - id: 00000000-0000-4000-8000-000000000015
                        match_id: 00000000-0000-4000-8000-000000000008
                        sequence: 1
                        minute: 1
                        type: MATCH_START
                        data:
                          home_score: 0
                          away_score: 0
        "404":
          description: Match not found in the caller's world.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: access_token
      description: httpOnly cookie set by POST /api/auth/login and rotated by /api/auth/refresh.
  parameters:
...
    MatchID:
      name: id
      in: path
      required: true
      description: Match id.
      schema: { type: string, format: uuid, example: "00000000-0000-4000-8000-000000000008" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: internal error }
...
    MatchEvent:
      type: object
      properties:
        id: { type: string, format: uuid }
        match_id: { type: string, format: uuid }
        sequence: { type: integer }
        minute: { type: integer }
        type: { type: string, example: MATCH_START }
        data:
          type: object
          additionalProperties: true
    Policy:
      type: object
      properties:
        world_id: { type: string, format: uuid }
        manager_id: { type: string, format: uuid }
        type: { type: string, enum: [squad, transfer, training] }
        params: { type: object, additionalProperties: true }
        enabled: { type: boolean }
        updated_at: { type: string, format: date-time }
    AbsenceView:
      type: object
      properties:
        manager_id: { type: string, format: uuid }
        world_id: { type: string, format: uuid }
        absence_state: { $ref: "#/components/schemas/AbsenceState" }
        club_ids:
          type: array
          items: { type: string, format: uuid }
        next_fixture_at: { type: string, format: date-time, nullable: true }
        delegation_active: { type: boolean }
    AbsenceState:
      type: object
      properties:
        away_since: { type: string, format: date-time, nullable: true }
        away_auto: { type: boolean }
        consecutive_missed: { type: integer }
        last_activity_at: { type: string, format: date-time, nullable: true }
...
security:
  - sessionCookie: []